Apperantly i forgot to add the test file for edges, although it doesn't really do much rn. Otherwise fixed a shit ton of shit like calloc arguments being swapped improperly initializing a node and other stuff which made the kruskal function not work. It works now for matrix and anyway tomorow/today is already another day

+++ C/Node.h
@@ -26,6 +26,6 @@ int compareToNode(Node * a, Node * b);
 void init_node(Node ** node, char * name, void (*printNode)(Node * n));
 void destroy_node(Node * n);
 void printNode(Node * n);
-void * putNeighbor(Node * a, Node * b, long weight);
+void * putNeighbor(Node * a, Node * b, int weight);
 
 #endif
